+++ simpletransformers/language_modeling/language_modeling_model.py
@@ -331,7 +331,8 @@ class LanguageModelingModel:
             eval_file (optional): Path to eval file containing the text to evaluate the language model on.
 
         Returns:
-            None
+            global_step: Number of global steps trained
+            training_details: Average training loss if evaluate_during_training is False or full training progress scores if evaluate_during_training is True
         """  # noqa: ignore flake8"
 
         if args:
@@ -361,7 +362,7 @@ class LanguageModelingModel:
 
         os.makedirs(output_dir, exist_ok=True)
 
-        global_step, tr_loss = self.train(
+        global_step, training_details = self.train(
             train_dataset,
             output_dir,
             show_running_loss=show_running_loss,
@@ -381,6 +382,8 @@ class LanguageModelingModel:
 
         if verbose:
             logger.info(" Training of {} model complete. Saved to {}.".format(self.args.model_type, output_dir))
+
+        return global_step, training_details
 
     def train(
         self, train_dataset, output_dir, show_running_loss=True, eval_file=None, verbose=True, **kwargs,
@@ -496,6 +499,7 @@ class LanguageModelingModel:
         logger.info(" Training started")
 
         global_step = 0
+        training_progress_scores = None
         tr_loss, logging_loss = 0.0, 0.0
         model.zero_grad()
         train_iterator = trange(int(args.num_train_epochs), desc="Epoch", disable=args.silent, mininterval=0)
@@ -694,7 +698,12 @@ class LanguageModelingModel:
                                             logger.info(f" Patience of {args.early_stopping_patience} steps reached.")
                                             logger.info(" Training terminated.")
                                             train_iterator.close()
-                                        return global_step, tr_loss / global_step
+                                        return (
+                                            global_step,
+                                            tr_loss / global_step
+                                            if not self.args.evaluate_during_training
+                                            else training_progress_scores,
+                                        )
                         else:
                             if results[args.early_stopping_metric] - best_eval_metric > args.early_stopping_delta:
                                 best_eval_metric = results[args.early_stopping_metric]
@@ -715,10 +724,18 @@ class LanguageModelingModel:
                                             logger.info(f" Patience of {args.early_stopping_patience} steps reached.")
                                             logger.info(" Training terminated.")
                                             train_iterator.close()
-                                        return global_step, tr_loss / global_step
+                                        return (
+                                            global_step,
+                                            tr_loss / global_step
+                                            if not self.args.evaluate_during_training
+                                            else training_progress_scores,
+                                        )
 
                 if args.max_steps > 0 and global_step > args.max_steps:
-                    return global_step, tr_loss / global_step
+                    return (
+                        global_step,
+                        tr_loss / global_step if not self.args.evaluate_during_training else training_progress_scores,
+                    )
 
             epoch_number += 1
             output_dir_current = os.path.join(output_dir, "checkpoint-{}-epoch-{}".format(global_step, epoch_number))
@@ -770,7 +787,12 @@ class LanguageModelingModel:
                                     logger.info(f" Patience of {args.early_stopping_patience} steps reached")
                                     logger.info(" Training terminated.")
                                     train_iterator.close()
-                                return global_step, tr_loss / global_step
+                                return (
+                                    global_step,
+                                    tr_loss / global_step
+                                    if not self.args.evaluate_during_training
+                                    else training_progress_scores,
+                                )
                 else:
                     if results[args.early_stopping_metric] - best_eval_metric > args.early_stopping_delta:
                         best_eval_metric = results[args.early_stopping_metric]
@@ -789,12 +811,23 @@ class LanguageModelingModel:
                                     logger.info(f" Patience of {args.early_stopping_patience} steps reached")
                                     logger.info(" Training terminated.")
                                     train_iterator.close()
-                                return global_step, tr_loss / global_step
+                                return (
+                                    global_step,
+                                    tr_loss / global_step
+                                    if not self.args.evaluate_during_training
+                                    else training_progress_scores,
+                                )
 
             if args.max_steps > 0 and global_step > args.max_steps:
-                return global_step, tr_loss / global_step
+                return (
+                    global_step,
+                    tr_loss / global_step if not self.args.evaluate_during_training else training_progress_scores,
+                )
 
-        return global_step, tr_loss / global_step
+        return (
+            global_step,
+            tr_loss / global_step if not self.args.evaluate_during_training else training_progress_scores,
+        )
 
     def eval_model(self, eval_file, output_dir=None, verbose=True, silent=False, **kwargs):
         """